requirements.txt
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.0
numpy==1.26.0
joblib==1.3.2
mlflow==2.8.0
pyspark==3.5.0
